requirements.txt
Sphinx==2.2.0
setuptools==41.4.0
pytest==5.2.1
python_dateutil==2.8.0
